LayoutTests/fast/security-labels/xhrsend-event-expected.txt

This tests that the checkxhrsend event behaves properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


checkxhrsend reached
PASS tag.isOn(event.securityLabel) is true
PASS req.send(tag.addTo("foo")); threw exception Error: XMLHttpRequest denied by checkxmlsend.
checkxhrsend reached
PASS tag.isOn(event.securityLabel) is false
PASS successfullyParsed is true

TEST COMPLETE
